+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Use run dir for cwd when calling bins
+* Moved body of Template to Genesis
 * Improved version task
 * Added signal handlers to bin launcher